requirements.txt
requests
pillows
